+++ mokoki-compiler/main/kotlin/com/javiersc/mokoki/compiler/MokokiCommandLineProcessor.kt
@@ -1,0 +1,13 @@
+package com.javiersc.mokoki.compiler
+
+import com.javiersc.mokoki.mokoki.compiler.MokokiCompilerProjectData
+import org.jetbrains.kotlin.compiler.plugin.AbstractCliOption
+import org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
+
+public class MokokiCommandLineProcessor : CommandLineProcessor {
+
+    override val pluginId: String =
+        "${MokokiCompilerProjectData.Group}.${MokokiCompilerProjectData.Name}"
+
+    override val pluginOptions: Collection<AbstractCliOption> = emptyList()
+}